pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "ffmpeg>=1.4",
    "ffmpeg-python>=0.2.0",
    "pillow>=11.3.0",
    "python-telegram-bot==20.8",
    "requests>=2.32.4",
]
